Trait ic_ic00_types::Payload

pub trait Payload<'a>: Sized + CandidType + Deserialize<'a> {
    fn encode(&self) -> Vec<u8> { ... }

    fn decode(blob: &'a [u8]) -> Result<Self, Error> { ... }
}

A trait to be implemented by all structs that are used as payloads by IC00. This trait encapsulates Candid serialization so that consumers of IC00 don't need to explicitly depend on Candid.

Provided methods

fn encode(&self) -> Vec<u8>

fn decode(blob: &'a [u8]) -> Result<Self, Error>

Loading content...